# Use rustc-hash in MIR optimizations