--- a/compiler/benches/benchmark.rs
+++ b/compiler/benches/benchmark.rs
@@ -2,13 +2,27 @@ use criterion::{
     criterion_group, criterion_main, measurement::Measurement, BatchSize, Bencher, BenchmarkId,
     Criterion,
 };
-use utils::{run, setup_and_compile};
+use utils::{compile, run, setup, setup_and_compile};
 
 mod utils;
 
 // `Core` is available via `use "..Core"`.
 
-fn vm_runtime(c: &mut Criterion) {
+fn benchmark_compiler(c: &mut Criterion) {
+    let mut group = c.benchmark_group("Compiler");
+
+    group.sample_size(100);
+    group.bench_function("hello_world", |b| {
+        b.compile(r#"main _ := "Hello, world!""#);
+    });
+
+    group.sample_size(20);
+    let fibonacci_code = create_fibonacci_code(15);
+    group.bench_function("fibonacci", |b| b.compile(&fibonacci_code));
+
+    group.finish();
+}
+fn benchmark_vm_runtime(c: &mut Criterion) {
     let mut group = c.benchmark_group("VM Runtime");
 
     group.sample_size(100);
@@ -18,7 +32,16 @@ fn vm_runtime(c: &mut Criterion) {
 
     group.sample_size(20);
     let n = 15;
-    let fibonacci_code = format!(
+    let fibonacci_code = create_fibonacci_code(n);
+    group.bench_function(BenchmarkId::new("fibonacci", n), |b| {
+        b.run_vm(&fibonacci_code)
+    });
+
+    group.finish();
+}
+
+fn create_fibonacci_code(n: usize) -> String {
+    format!(
         r#"[ifElse, int] = use "..Core"
 
 fibRec = {{ fibRec n ->
@@ -31,19 +54,22 @@ fib n =
   needs (int.is n)
   fibRec fibRec n
 
-main _ := fib {n}"#
-    );
-    group.bench_function(BenchmarkId::new("fibonacci", n), |b| {
-        b.run_vm(&fibonacci_code)
-    });
-
-    group.finish();
+main _ := fib {n}"#,
+    )
 }
 
 trait BencherExtension {
+    fn compile(&mut self, source_code: &str);
     fn run_vm(&mut self, source_code: &str);
 }
 impl<'a, M: Measurement> BencherExtension for Bencher<'a, M> {
+    fn compile(&mut self, source_code: &str) {
+        self.iter_batched(
+            || setup(),
+            |mut db| compile(&mut db, source_code),
+            BatchSize::SmallInput,
+        )
+    }
     fn run_vm(&mut self, source_code: &str) {
         self.iter_batched(
             || setup_and_compile(source_code),
@@ -53,5 +79,5 @@ impl<'a, M: Measurement> BencherExtension for Bencher<'a, M> {
     }
 }
 
-criterion_group!(benches, vm_runtime);
+criterion_group!(benches, benchmark_compiler, benchmark_vm_runtime);
 criterion_main!(benches);