<!-- Please enter the corresponding issue ID: -->
Closes: #270

<!-- Please summarize your changes: -->
https://crates.io/crates/rustc-hash

Baseline benchmark results for `fibonacci 15` (`hyperfine "cargo run --release --manifest-path=compiler/Cargo.toml -- run packages/examples/fibonacci.candy"`):

```
Time (mean ± σ):      1.056 s ±  0.057 s    [User: 0.973 s, System: 0.070 s]
Range (min … max):    0.994 s …  1.137 s    10 runs
```

`FxHasher` in `Mir::do_hash`:

```
Time (mean ± σ):      1.062 s ±  0.165 s    [User: 0.961 s, System: 0.068 s]
Range (min … max):    0.957 s …  1.515 s    10 runs
```

`FxHashMap` in `Mir::eliminate_common_subtrees`:

```
Time (mean ± σ):     932.4 ms ±  61.4 ms    [User: 852.3 ms, System: 59.5 ms]
Range (min … max):   868.8 ms … 1073.8 ms    10 runs
```

`FxHashMap`/`FxHashSet` in remaining MIR optimizations:

```
Time (mean ± σ):     830.1 ms ±  23.8 ms    [User: 763.7 ms, System: 62.5 ms]
Range (min … max):   794.1 ms … 860.3 ms    10 runs
```

---

New baseline (benchmarks were getting a bit flaky and slower):

```
Time (mean ± σ):      1.036 s ±  0.025 s    [User: 0.955 s, System: 0.076 s]
Range (min … max):    0.981 s …  1.068 s    10 runs
```

`FxHashMap` in the VM:

```
Time (mean ± σ):     899.2 ms ±  62.8 ms    [User: 829.7 ms, System: 64.3 ms]
Range (min … max):   838.5 ms … 1022.1 ms    10 runs
```

```
# `cargo bench`
VM Runtime/hello_world  time:   [10.920 ms 11.159 ms 11.435 ms]
                        change: [-11.848% -9.3501% -6.5315%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe

VM Runtime/fibonacci/15 time:   [155.88 ms 162.01 ms 171.53 ms]
                        change: [-66.391% -64.448% -61.775%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 20 measurements (15.00%)
  2 (10.00%) high mild
  1 (5.00%) high severe
```

### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)